chore: organize types
