+++ src/node/AttributeInputNode.ts
@@ -1,5 +1,6 @@
 import {AttributeInputNodeData} from '../types/CommonType';
-import Node, {NodeClassNames} from './Node';
+import {NodeClassNames} from './INode';
+import Node from './Node';
 
 export default class AttributeInputNode extends Node {
   protected __nodeData: AttributeInputNodeData;
