+++ src/node/VaryingInputNode.ts
@@ -1,5 +1,6 @@
 import {VaryingInputNodeData} from '../types/CommonType';
-import Node, {NodeClassNames} from './Node';
+import {NodeClassNames} from './INode';
+import Node from './Node';
 
 export default class VaryingInputNode extends Node {
   protected __nodeData: VaryingInputNodeData;
