+++ src/types/CommonType.ts
@@ -23,23 +23,6 @@ export type ShaderConstantValueObject = _ShaderConstantValueObject;
 export type ShaderUniformObject = _ShaderUniformObject;
 export type ShaderVaryingObject = _ShaderVaryingObject;
 
-export type GlslTypeStr =
-  | 'bool'
-  | 'int'
-  | 'float'
-  | 'vec2'
-  | 'vec3'
-  | 'vec4'
-  | 'ivec2'
-  | 'ivec3'
-  | 'ivec4'
-  | 'sampler2D'
-  | 'samplerCube'
-  | 'mat2'
-  | 'mat3'
-  | 'mat4'
-  | 'unknown';
-
 export interface ShaderCodes {
   vertexShader: string;
   fragmentShader: string;
